Move IDObject class definition to a new file and update its implementation
